# Refactor await and yield implementation and enable async iteration