--- a/lib/Backend/IR.cpp
+++ b/lib/Backend/IR.cpp
@@ -1049,8 +1049,7 @@ bool IR::Instr::IsStElemVariant() const
 
 bool IR::Instr::DontHoistBailOnNoProfileAboveInGeneratorFunction() const
 {
-    return this->m_opcode == Js::OpCode::ResumeYield ||
-        this->m_opcode == Js::OpCode::ResumeYieldStar ||
+    return this->m_opcode == Js::OpCode::GeneratorResumeYield ||
         this->m_opcode == Js::OpCode::GeneratorCreateInterpreterStackFrame;
 }