--- a/lib/Backend/IRBuilder.cpp
+++ b/lib/Backend/IRBuilder.cpp
@@ -1864,39 +1864,6 @@ IRBuilder::BuildReg2(Js::OpCode newOpcode, uint32 offset, Js::RegSlot R0, Js::Re
         dstOpnd->SetValueType(ValueType::String);
         break;
 
-    case Js::OpCode::ResumeYield:
-    {
-        IR::Instr* loadResumeYieldData = IR::Instr::New(Js::OpCode::GeneratorLoadResumeYieldData, src1Opnd /* dst */, m_func);
-        this->AddInstr(loadResumeYieldData, offset);
-
-        // Insert bailout for debugger, since we are bailing out to the ResumeYield instruction (OP_ResumeYield) in the interpreter,
-        // we have to load the ResumeYieldData first
-        if (this->m_func->IsJitInDebugMode())
-        {
-            this->InsertBailOutForDebugger(offset, IR::BailOutForceByFlag | IR::BailOutBreakPointInFunction | IR::BailOutStep);
-        }
-
-        break;
-    }
-
-    case Js::OpCode::Await:
-    case Js::OpCode::AsyncYield:
-    case Js::OpCode::AsyncYieldStar:
-    {
-        // All of these opcodes rely on the ResumeYieldData passed as an argument to the jit'd frame,
-        // load it first before we do the actual operation.
-        // Also need to create the instruction differently because, unlike other Reg2 instructions,
-        // these opcodes actually use their destination operand as a source and have no destination.
-
-        IR::Instr* loadResumeYieldData = IR::Instr::New(Js::OpCode::GeneratorLoadResumeYieldData, dstOpnd /* dst */, m_func);
-        this->AddInstr(loadResumeYieldData, offset);
-
-        instr = IR::Instr::New(newOpcode, nullptr /* dst */, dstOpnd /* src1 */, src1Opnd /* src2 */, m_func);
-        this->AddInstr(instr, offset);
-
-        return;
-    }
-
     case Js::OpCode::Yield:
         instr = IR::Instr::New(newOpcode, dstOpnd, src1Opnd, m_func);
         this->AddInstr(instr, offset);
@@ -1925,13 +1892,13 @@ IRBuilder::BuildReg2(Js::OpCode newOpcode, uint32 offset, Js::RegSlot R0, Js::Re
         }
 #endif
 
-        // This label indicates the section where we start loading the ResumeYieldData on the stack
-        // that comes from either .next(), .return(), or .throw() to the right symbol and finally
-        // extract its data through Op_ResumeYield
-        IR::LabelInstr* resumptionLabel = IR::LabelInstr::New(Js::OpCode::GeneratorResumeYieldLabel, m_func);
-        resumptionLabel->m_hasNonBranchRef = true;          // set to true so that we don't move this label around
-        LABELNAMESET(resumptionLabel, "ResumeYieldHelperLabel");
-        this->AddInstr(resumptionLabel, offset);
+        IR::Instr* resumeYield = IR::Instr::New(Js::OpCode::GeneratorResumeYield, dstOpnd, m_func);
+        this->AddInstr(resumeYield, offset);
+
+        if (this->m_func->IsJitInDebugMode())
+        {
+            this->InsertBailOutForDebugger(offset, IR::BailOutForceByFlag | IR::BailOutBreakPointInFunction | IR::BailOutStep);
+        }
 
         return;
     }
@@ -2153,19 +2120,6 @@ IRBuilder::BuildReg3(Js::OpCode newOpcode, uint32 offset, Js::RegSlot dstRegSlot
     else
     {
         instr = IR::Instr::New(newOpcode, dstOpnd, src1Opnd, src2Opnd, m_func);
-    }
-
-    if (newOpcode == Js::OpCode::ResumeYieldStar)
-    {
-        IR::Instr* loadResumeYieldData = IR::Instr::New(Js::OpCode::GeneratorLoadResumeYieldData, src1Opnd /* dst */, m_func);
-        this->AddInstr(loadResumeYieldData, offset);
-
-        // Insert bailout for debugger, since we are bailing out to the ResumeYieldStar instruction (OP_ResumeYield) in the interpreter,
-        // we have to load the ResumeYieldData first
-        if (this->m_func->IsJitInDebugMode())
-        {
-            this->InsertBailOutForDebugger(offset, IR::BailOutForceByFlag | IR::BailOutBreakPointInFunction | IR::BailOutStep);
-        }
     }
 
     this->AddInstr(instr, offset);