--- a/lib/Backend/Lower.cpp
+++ b/lib/Backend/Lower.cpp
@@ -2962,44 +2962,12 @@ Lowerer::LowerRange(IR::Instr *instrStart, IR::Instr *instrEnd, bool defaultDoFa
             break;
         }
 
-        case Js::OpCode::Await:
+        case Js::OpCode::NewAwaitObject:
         {
-            IR::Opnd *srcOpnd1 = instr->UnlinkSrc1();
-            IR::Opnd *srcOpnd2 = instr->UnlinkSrc2();
+            IR::Opnd *src1Opnd = instr->UnlinkSrc1();
             LoadScriptContext(instr);
-            m_lowererMD.LoadHelperArgument(instr, srcOpnd2);
-            m_lowererMD.LoadHelperArgument(instr, srcOpnd1);
-            m_lowererMD.ChangeToHelperCall(instr, IR::HelperAwait);
-            break;
-        }
-
-        case Js::OpCode::AsyncYield:
-        {
-            IR::Opnd *srcOpnd1 = instr->UnlinkSrc1();
-            IR::Opnd *srcOpnd2 = instr->UnlinkSrc2();
-            LoadScriptContext(instr);
-            m_lowererMD.LoadHelperArgument(instr, srcOpnd2);
-            m_lowererMD.LoadHelperArgument(instr, srcOpnd1);
-            m_lowererMD.ChangeToHelperCall(instr, IR::HelperAsyncYield);
-            break;
-        }
-
-        case Js::OpCode::AsyncYieldIsReturn:
-        {
-            IR::Opnd *srcOpnd1 = instr->UnlinkSrc1();
-            m_lowererMD.LoadHelperArgument(instr, srcOpnd1);
-            m_lowererMD.ChangeToHelperCall(instr, IR::HelperAsyncYieldIsReturn);
-            break;
-        }
-
-        case Js::OpCode::AsyncYieldStar:
-        {
-            IR::Opnd *srcOpnd1 = instr->UnlinkSrc1();
-            IR::Opnd *srcOpnd2 = instr->UnlinkSrc2();
-            LoadScriptContext(instr);
-            m_lowererMD.LoadHelperArgument(instr, srcOpnd2);
-            m_lowererMD.LoadHelperArgument(instr, srcOpnd1);
-            m_lowererMD.ChangeToHelperCall(instr, IR::HelperAsyncYieldStar);
+            m_lowererMD.LoadHelperArgument(instr, src1Opnd);
+            m_lowererMD.ChangeToHelperCall(instr, IR::HelperNewAwaitObject);
             break;
         }
 
@@ -3009,16 +2977,9 @@ Lowerer::LowerRange(IR::Instr *instrStart, IR::Instr *instrEnd, bool defaultDoFa
             break;
         }
 
-        case Js::OpCode::GeneratorLoadResumeYieldData:
+        case Js::OpCode::GeneratorResumeYield:
         {
-            this->m_lowerGeneratorHelper.LowerGeneratorLoadResumeYieldData(instr);
-            break;
-        }
-
-        case Js::OpCode::ResumeYield:
-        case Js::OpCode::ResumeYieldStar:
-        {
-            this->m_lowerGeneratorHelper.LowerResumeGenerator(instr);
+            this->m_lowerGeneratorHelper.LowerGeneratorResumeYield(instr);
             break;
         }
 
@@ -3148,7 +3109,6 @@ Lowerer::LowerRange(IR::Instr *instrStart, IR::Instr *instrEnd, bool defaultDoFa
         case Js::OpCode::GeneratorOutputBailInTraceLabel:
 #endif
         case Js::OpCode::GeneratorBailInLabel:
-        case Js::OpCode::GeneratorResumeYieldLabel:
         case Js::OpCode::GeneratorEpilogueFrameNullOutLabel:
         case Js::OpCode::GeneratorEpilogueNoFrameNullOutLabel:
             Assert(this->m_func->GetJITFunctionBody()->IsCoroutine());
@@ -26552,7 +26512,6 @@ Lowerer::ValidOpcodeAfterLower(IR::Instr* instr, Func * func)
     case Js::OpCode::GeneratorOutputBailInTraceLabel:
 #endif
     case Js::OpCode::GeneratorBailInLabel:
-    case Js::OpCode::GeneratorResumeYieldLabel:
     case Js::OpCode::GeneratorEpilogueFrameNullOutLabel:
     case Js::OpCode::GeneratorEpilogueNoFrameNullOutLabel:
         return func->GetJITFunctionBody()->IsCoroutine();
@@ -29303,7 +29262,7 @@ Lowerer::LowerGeneratorHelper::LowerGeneratorTraceBailIn(IR::Instr* instr)
 #endif
 
 IR::SymOpnd*
-Lowerer::LowerGeneratorHelper::CreateResumeYieldDataOpnd() const
+Lowerer::LowerGeneratorHelper::CreateResumeYieldOpnd() const
 {
     StackSym* resumeYieldDataSym = StackSym::NewImplicitParamSym(4, this->func);
     this->func->SetArgOffset(resumeYieldDataSym, (LowererMD::GetFormalParamOffset() + 1) * MachPtr);
@@ -29311,23 +29270,13 @@ Lowerer::LowerGeneratorHelper::CreateResumeYieldDataOpnd() const
 }
 
 void
-Lowerer::LowerGeneratorHelper::LowerGeneratorLoadResumeYieldData(IR::Instr* instr)
+Lowerer::LowerGeneratorHelper::LowerGeneratorResumeYield(IR::Instr* instr)
 {
-    // prm2 is the ResumeYieldData pointer per calling convention established in JavascriptGenerator::CallGenerator
+    // prm2 is the resume yield object var per calling convention established in JavascriptGenerator::CallGenerator
     // This is the value the bytecode expects to be in the dst register of the Yield opcode after resumption.
     // Load it here after the bail-in.
-    this->lowerer->InsertMove(instr->UnlinkDst(), this->CreateResumeYieldDataOpnd(), instr);
+    this->lowerer->InsertMove(instr->UnlinkDst(), this->CreateResumeYieldOpnd(), instr);
     instr->Unlink();
-}
-
-void
-Lowerer::LowerGeneratorHelper::LowerResumeGenerator(IR::Instr* instr)
-{
-    IR::Opnd* srcOpnd1 = instr->UnlinkSrc1();
-    IR::Opnd* srcOpnd2 = instr->m_opcode == Js::OpCode::ResumeYieldStar ? instr->UnlinkSrc2() : IR::AddrOpnd::NewNull(this->func);
-    this->lowererMD.LoadHelperArgument(instr, srcOpnd2);
-    this->lowererMD.LoadHelperArgument(instr, srcOpnd1);
-    this->lowererMD.ChangeToHelperCall(instr, IR::HelperResumeYield);
 }
 
 void