--- a/lib/Common/BackendApi.h
+++ b/lib/Common/BackendApi.h
@@ -276,6 +276,7 @@ enum VTableValue {
     VtableVirtualJavascriptGeneratorFunctionWithComputedName,
     VtableVirtualJavascriptGeneratorFunctionWithHomeObjAndComputedName,
     VtableJavascriptAsyncFunction,
+    VtableJavascriptAsyncGeneratorFunction,
 
     VtableStackScriptFunction,
     VtableAsmJsScriptFunctionWithComputedName,