--- a/lib/Jsrt/JsrtDebugUtils.cpp
+++ b/lib/Jsrt/JsrtDebugUtils.cpp
@@ -271,6 +271,7 @@ void JsrtDebugUtils::AddPropertyType(Js::DynamicObject * object, Js::IDiagObject
         case Js::TypeIds_WeakSet:
         case Js::TypeIds_SymbolObject:
         case Js::TypeIds_Generator:
+        case Js::TypeIds_AsyncGenerator:
         case Js::TypeIds_AsyncFromSyncIterator:
         case Js::TypeIds_Promise:
         case Js::TypeIds_GlobalObject:
@@ -472,6 +473,7 @@ const char16 * JsrtDebugUtils::GetClassName(Js::TypeId typeId)
     case Js::TypeIds_WeakSet:               return _u("WeakSet");
     case Js::TypeIds_SymbolObject:          return _u("Symbol");
     case Js::TypeIds_Generator:             return _u("Generator");
+    case Js::TypeIds_AsyncGenerator:        return _u("AsyncGenerator");
     case Js::TypeIds_AsyncFromSyncIterator: return _u("AsyncFromSyncIterator");
     case Js::TypeIds_Promise:               return _u("Promise");
     case Js::TypeIds_GlobalObject:          return _u("Object");