--- a/lib/Parser/Parse.cpp
+++ b/lib/Parser/Parse.cpp
@@ -7328,7 +7328,6 @@ void Parser::ParseExpressionLambdaBody(ParseNodeFnc * pnodeLambda, bool fAllowIn
         // Pushing a statement node with PushStmt<>() normally does this initialization
         // but do it here manually since we know there is no outer statement node.
         pnodeRet->grfnop = 0;
-        pnodeRet->pnodeOuter = nullptr;
 
         pnodeLambda->ichLim = max(pnodeRet->ichLim, lastRParen);
         pnodeLambda->cbLim = this->GetScanner()->IecpLimTokPrevious();
@@ -9748,7 +9747,6 @@ ParseNodeStmt * Parser::ParseTryCatchFinally()
         if (buildAST)
         {
             pnodeTC = CreateNodeForOpT<knopTryCatch>();
-            pnodeT->pnodeOuter = pnodeTC;
             pnodeTC->pnodeTry = pnodeT;
         }
         PushStmt<buildAST>(&stmt, pnodeTC, knopTryCatch, nullptr);
@@ -9783,14 +9781,11 @@ ParseNodeStmt * Parser::ParseTryCatchFinally()
         if (!hasCatch)
         {
             pnodeTF->pnodeTry = pnodeT;
-            pnodeT->pnodeOuter = pnodeTF;
         }
         else
         {
             pnodeTF->pnodeTry = CreateNodeForOpT<knopTry>();
-            pnodeTF->pnodeTry->pnodeOuter = pnodeTF;
             pnodeTF->pnodeTry->pnodeBody = pnodeTC;
-            pnodeTC->pnodeOuter = pnodeTF->pnodeTry;
         }
         pnodeTF->pnodeFinally = pnodeFinally;
     }