--- a/lib/Runtime/ByteCode/ByteCodeCacheReleaseFileVersion.h
+++ b/lib/Runtime/ByteCode/ByteCodeCacheReleaseFileVersion.h
@@ -5,6 +5,6 @@
 // NOTE: If there is a merge conflict the correct fix is to make a new GUID.
 // This file was generated with tools\update_bytecode_version.ps1
 
-// {5C20FA9E-A347-4731-A0DE-148B4FA12B04}
+// {4CD08EFA-A1E7-4C2D-81CE-D70DB9D553E2}
 const GUID byteCodeCacheReleaseFileVersion =
-{ 0x5C20FA9E, 0xA347, 0x4731, { 0xA0, 0xDE, 0x14, 0x8B, 0x4F, 0xA1, 0x2B, 0x04 } };
+{ 0x4CD08EFA, 0xA1E7, 0x4C2D, { 0x81, 0xCE, 0xD7, 0x0D, 0xB9, 0xD5, 0x53, 0xE2 } };