--- a/lib/Runtime/Debug/TTSupport.h
+++ b/lib/Runtime/Debug/TTSupport.h
@@ -255,8 +255,8 @@ namespace TTD
             SnapGeneratorVirtualScriptFunction,
             SnapAsyncFunction,
             SnapGenerator,
-            JavascriptPromiseAsyncSpawnExecutorFunction,
-            JavascriptPromiseAsyncSpawnStepArgumentExecutorFunction,
+            JavascriptAsyncSpawnExecutorFunction,
+            JavascriptAsyncSpawnStepFunction,
 
             //objects that should always be well known but which may have other info we want to restore
             SnapWellKnownObject,