## Overview

The goals of this change are to:

- Enable async generators
- Streamline and simplify the relationship between generators, async functions, and async generators.
- Make async functions and generators more JIT-friendly by reducing the number and complexity of supporting helpers.
- Improve code readability and maintainability.

The primary architectural change is to perform yield resume logic and branching directly in bytecode, rather than relying on the `ResumeYield` helper, and to desugar all async functions and async generators to normal generators. This allows us to eliminate the special-case `GeneratorReturnException` exception type.

In order to perform yield resume logic in bytecode, we resume the generator by creating a dynamic object with "value" (an arbitrary var) and "kind" (tagged int) fields. Branch instructions are inserted to deal with the various "kind" cases: "normal", "return", and "throw".

To desugar async functions and async generators to normal generators, we introduce a new type ID to represent "await results". Async generators then test for this type ID when determining whether their internal coroutine is "awaiting" or "yielding".

## Changes

### Architectural

- Remove `ResumeYieldData`. Instead, resume the generator with a "resume yield object" containing "value" and "kind" properties.
- Remove `GeneratorReturnException`. The bytecode is now responsible for returning based on the resume yield kind.
- Remove opcodes `ResumeYield`, `Await`, `AsyncYield`, `AsyncYieldStar`, `AsyncYieldIsReturn`.
- Add new opcode and helper for `NewAwaitObject`. This object type is yielded when awaiting and is uniquely distinguished by type ID from "normal" yield results.
- Remove `ResumeYield` and `ResumeYieldStar` opcodes and helpers. When generators are resumed, the resume yield object is loaded into the yield register and the resume logic is performed in bytecode.
- Make async generators a subclass of `JavascriptGenerator`.
- Enable flag `ES2018AsyncIteration` by default.

### JIT

- Remove unused `GeneratorResumeYieldLabel`.
- Rename `GeneratorLoadResumeYieldData` to `GeneratorResumeYield`.
- Remove support for `ResumeYield`. `GeneratorResumeYield` is inserted immediately after the bail-in instr and takes care of loading the resume yield object.

### ByteCode

- Rename `EmitDummyYield` to `EmitStartupYield`.
- Remove unnecessary wrapping of async function body in try/catch. Exception handling is performed by library code that wraps the coroutine.
- Move startup yield after function declaration hoisting to unify module and generator code paths.
- Implement iterator next method caching to match spec.
- Replace `tryScopeRecordsList` and loop "parent node linking" with a single mechanism for performing cleanup in both the branch-in-loop cases and yield cases.
- Implement yield star evaluation logic in bytecode.
- Implement yield resume branching in bytecode.
- Await `return` operand in async generators to match spec.
- Add new error message for when `yield *` inner iterator does not have a "throw" method.

### Parser

- Remove `ParseNodeLoop`. The only purpose of this node type was to store an outer statement reference used when emitting jump cleanup. This information is now maintained in the bytecode generator's `jumpCleanupList`.

### Library

- Move `JavascriptAsyncGeneratorFunction` into a new file.
- Add type ID for async generators.
- Remove unnecessary `JavascriptPromiseAsyncSpawnExecutorFunction`.
- Use more readable names for `CreateIteratorResultObject` functions.
- Use more readable names for async function support classes.
- Move `JavascriptAsyncFunction` into a new file (previously code was split between `JavascriptPromise` and `JavascriptGeneratorFunction`).
- Move `JavascriptAsyncGeneratorFunction` into a new file.
- Create locked library types for await objects and resume yield objects.
- Move "copy arguments to heap" logic into `JavascriptGenerator::New`.

## Performance

This change results in a small performance increase on the ARES-6 benchmark.

| Scenario | Average Time | Change over baseline |
|----------|--------------|----------------------|
| ARES-6/Basic (No JIT) |	3671ms | 3.96% faster |
| ARES-6/Basic (JIT) | 2502ms | 2.56% faster |