@rhuanjl 2019-10-28
It's nice to see the horrible OP_ResumeYield logic replaced.

A little sad to see that much of my implementation of AsyncGenerators is removed/re-written here BUT I can see that this way of doing it is a lot neater.

A couple of things stand out to me by way of future improvements:
 - The startup yield is only needed when the generator has complex parameters - there was code to put a relevant condition around it in the PR to enable jitting generators - obviously that PR as it stands is dead with this change but that enhancement seems worth bringing across to me.
 - In my asyncGenerator implementation I cached the continuation methods for the async generator rather than creating them every time obviously doing that has some memory cost but I figure it should be worth it for performance and the same could/should be done for async functions.
 - Similarly the await object should only need to be created once and could be cached in an internal slot
 - The caching of the iterator Next property is missing in some other places not just generators