@zenparsing 2019-10-28
Hi @rhuanjl - I was a bit nervous about posting this PR given there were so many changes to what you had done, which was awesome work given the existing generators implementation. But particularly with JITing coming I want to take this opportunity to clean up the existing oddness. Apologies!

> In my asyncGenerator implementation I cached the continuation methods for the async generator rather than creating them every time obviously doing that has some memory cost but I figure it should be worth it for performance and the same could/should be done for async functions.

Oh right, I had forgotten about that, and it was a good idea. Hopefully this change will make it a bit easier to apply the same optimization to async functions (since there are fewer continuation functions to worry about).

> Similarly the await object should only need to be created once and could be cached in an internal slot

Yes! I was trying to figure out whether it would be better to store it in a C++ field or create a register for it in bytecode, but I didn't try to implement it either way yet.

> The caching of the iterator Next property is missing in some other places not just generators

Right. My idea was to go ahead and take care of the bytecode part here, and then use your existing PR for the remaining changes to the library code. I *think* I got all of the bytecode here?