@zenparsing 2019-10-29
Those changes look great! We can bring them in after this one gets landed.

For the TTD stuff, I don't really know it either but I think @boingoing could help when we get there. I'm also trying to consider what we want to do with TTD longer-term.