@fatcerberus 2019-11-04
> Remove unnecessary wrapping of async function body in try/catch. Exception handling is performed by library code that wraps the coroutine.

Ooh! Will that fix #4630, I wonder...? 🤞