@rhuanjl 2019-11-04
> > Remove unnecessary wrapping of async function body in try/catch. Exception handling is performed by library code that wraps the coroutine.
> 
> Ooh! Will that fix #4630, I wonder...? 🤞

It doesn't - BUT it lays the groundwork for an easy fix of #4630 that I have ready to submit after this lands (if someone doesn't beat me to it).