@pleath 2019-12-03
Note that the *_ReuseLoc variants should not be used the first time a code sequence writes to a register, only the 2nd and subsequent times. So:
R3 = LdFld R4.x
R3 = LdFld_ReuseLoc R5.x
Not:
R3 = LdFld_ReuseLoc R4.x
R3 = LdFld_ReuseLoc R5.x

The reason is that _ReuseLoc tells the JIT not to start a new lifetime. So in the above example, each _ReuseLoc op extends an existing lifetime of R3.