@zenparsing 2020-04-10
@rhuanjl This PR was already updated per @pleath's comments above. I'll try to rebase, regenerate bytecode, and give it another review within a couple of days.