@rhuanjl 2020-05-14
@zenparsing Please could you rebase this and regen the bytecode? I'll give it another review - I think I'm familiar enough now with most of this logic.

(I've got some follow up PRs I want to open)