@rhuanjl 2020-06-04
Merged as https://github.com/microsoft/ChakraCore/commit/c848d4d8d50c0dfb4a23540a9ee6cd023fa029c1